requirements.txt
pandas
numpy
yfinance
requests
scikit-learn
scipy
matplotlib
seaborn
pyyaml
jupyter
openmeteo-requests
requests-cache
retry_requests